feat:针对消息增加外挂状态数据,完成初期消息已读状态实现.fix:漫游去重逻辑未生效.
