+++ src/IM/listener/index.js
@@ -3,6 +3,7 @@ import { imReviceMessageListener } from './imReciveMessageListener'
 import { imPresenceListener } from './imPresenceListener'
 import { imContactListener } from './imContactListener'
 import { imGroupListener } from './imGroupListener'
+import { imReadAckListener } from './imReadAckListener'
 /* mount all listener */
 export const mountAllEMListener = () => {
     const { mountConnectEventListener } = imConnectListener()
@@ -15,11 +16,14 @@ export const mountAllEMListener = () => {
     mountContactEventListener()
     const { mountGroupEventListener } = imGroupListener()
     mountGroupEventListener()
+    const { mountReadAckEventListener } = imReadAckListener()
+    mountReadAckEventListener()
 }
 export {
     imConnectListener,
     imPresenceListener,
     imReviceMessageListener,
     imContactListener,
-    imGroupListener
+    imGroupListener,
+    imReadAckListener
 }
