+++ src/constant/messageType.js
@@ -1,14 +1,14 @@
-const SESSION_MESSAGE_TYPE = {
+export const SESSION_MESSAGE_TYPE = {
     img: '[图片]',
     file: '[文件]',
     audio: '[语音]',
     loc: '[位置]'
 }
 
-const CUSTOM_TYPE = {
+export const CUSTOM_TYPE = {
     userCard: '个人名片'
 }
-const ALL_MESSAGE_TYPE = {
+export const ALL_MESSAGE_TYPE = {
     TEXT: 'txt',
     IMAGE: 'img',
     AUDIO: 'audio',
@@ -19,25 +19,41 @@ const ALL_MESSAGE_TYPE = {
     CMD: 'cmd',
     INFORM: 'inform' //这个类型不在环信消息类型内，属于自己定义的一种系统通知类的消息。
 }
-const CHAT_TYPE = {
+export const CHAT_TYPE = {
     SINGLE: 'singleChat',
     GROUP: 'groupChat'
 }
 
-const MENTION_ALL = {
+export const MENTION_ALL = {
     TEXT: '所有人',
     VALUE: 'ALL'
 }
-const CHANGE_MESSAGE_BODAY_TYPE = {
+export const CHANGE_MESSAGE_BODAY_TYPE = {
     RECALL: 0,
     DELETE: 1,
     MODIFY: 2
 }
+
+export const MESSAGE_STATUS = {
+    SUCCESS: 'success',
+    FAIL: 'fail',
+    SENDING: 'sending'
+}
+export const MESSAGE_STATUS_TYPE = {
+    READ_STATUS: 'readStatus',
+    SEND_STATUS: 'sendStatus',
+    CHANLE_STATUS: 'chanleStatus'
+}
+
+//单个消息列表最大长度
+export const MAX_MESSAGE_LIST_COUNT = 100
 export default {
     SESSION_MESSAGE_TYPE,
     CUSTOM_TYPE,
     ALL_MESSAGE_TYPE,
     CHAT_TYPE,
     MENTION_ALL,
-    CHANGE_MESSAGE_BODAY_TYPE
+    CHANGE_MESSAGE_BODAY_TYPE,
+    MESSAGE_STATUS,
+    MESSAGE_STATUS_TYPE
 }
